LayoutTests/platform/ios-sim/fast/dom/custom/lifecycle-ready-creation-api-expected.txt

CONSOLE MESSAGE: line 14: TypeError: undefined is not a function (evaluating 'document.register("x-foo", { prototype: Object.create(HTMLElement.prototype, { readyCallback: { value: markingReadyCallback } }) })')
This test ensures that the lifecycle callbacks of API-originated elements are visible in following script block.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL document.createElement('x-foo').callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL document.createElement('div', 'x-bar').callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL document.createElementNS('http://www.w3.org/1999/xhtml', 'x-foo').callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL document.createElementNS('http://www.w3.org/1999/xhtml', 'div', 'x-bar').callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL foo.cloneNode().callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL foo.cloneNode(true).firstChild.callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL bar.cloneNode().callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL bar.cloneNode(true).firstChild.callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
PASS foreignFoo.callbacksCalled is undefined.
PASS foreignFoo.firstChild.callbacksCalled is undefined.
FAIL importedFoo.callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL importedFoo.firstChild.callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
PASS window.callbacksCalled is []
FAIL importedBar.callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL importedBar.firstChild.callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL window.callbacksCalled should be DIV,X-FOO. Was .
FAIL window.callbacksCalled should be X-FOO. Was .
FAIL window.callbacksCalled should be X-FOO. Was .
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
